Update Linq2GraphQL client generator with enum strategy support
A notable change in the client generator added a strategy for enums, which includes the option of mapping unknown values to a distinct member through command line option. Data type of quantity fields in different classes is also changed from float to double. The version of runtime in EnumTemplate is updated.